Fetch/pull git repo to depth of 1
The current static site HTML backup is around 1GB. We've had a number of
stalled builds lately where it seems to be getting stuck at this stage,
and it's probably because Heroku doesn't like us using this much space
on their ephemeral filesystem. By only pulling the latest commit, we
reduce our footprint to about 50MB - which should be fine.