Publish all unpublished measure versions
We want all of our measure versions to be available, even if
historically we published versions with errors. This promotes
transparency and is a good statistical practice. It also means that
users can link explicitly to the version of the data/commentary that
they saw on a specific date via permalinks.

 ## Ticket
https://trello.com/c/Vksqpdrm