Add question to select updated version
When a measure version update is correcting a factual mistake, the user
has to specifically note that this is the case. Where it is correcting a
fact, it's also useful to know how long the mistake has been published
(i.e. which version introduced the mistake). This patch adds a radio
question allowing the user to pick the earliest version that included
the error. This allows us to only show a banner declaring a factual
error for the real duration of the error.

 ## Ticket
https://trello.com/c/05vv68FM