Publish all versions (#1112)
This updates the static build process to publish ALL approved versions of each measure, rather than just the most recent one. This means, for example, that versions 1.0, 1.1 and 2.0 are all public, rather than 1.1 and 2.0. The 'latest' measure (eg 2.0 in this example) is published under two URLs, a versioned on (/2.0) and a /latest one.