Set external edit summary for all 1.0 measures
When we last ran this migration, we had an erroneous 'published IS TRUE'
condition on this update. We had one legitimate measure that therefore
didn't get the external edit summary. Come time to publish this (2 years
later!), it now can't be submitted for review due to this field being
required.