Add checkboxes to admin data source search
We currently have a page that allows an admin to see all data sources
entered into the system. We want the user to be able to manage the list
of data sources (by combining multiple ones into a single authoritative
source), so we need a way for users to select which data sources they
wish to combine.

This patch adds checkboxes to each data source in the list, allowing the
user to select which they wish to manage. The form currently does a GET
back to its own URL, but will soon be updated to point at the "merge
data sources" page, when it becomes available.

 ## Ticket
https://trello.com/c/w471ZapE