Remove ordering from data_sources relation
It doesn't make sense to order the data sources for a measure version by
the data source id, as we really want to keep track of the order that
data sources were linked to the measure version. We don't have an
explicit way to do this at the moment, as the join table doesn't have
its own incremental ID. Removing the explicit ordering will give us a
Postgres internal order, which will generally be the order the rows were
added the table, which works for us.