Allow a data source to be selected
This updates the searchable list of data sources with radio buttons for
each source, allowing one to be selected by the user as the source to
add to their measure version. It also adds an option for a user to
create a new data source, if they can't find one that matches.