Dynamic back link when creating data source
In the new flow, you can only get to the 'create a data source' page
after searching for existing data sources. You might click 'create a
data source' by accident, and then expect to click the back link to go
back to your search results and select the correct one. This updates
enables that behaviour by passing the searched-for term into the create
data source form. If it's available, the back link takes you back to
search results - if it's not (e.g. after successfully creating the data
source), it instead takes you back to the edit measure page.