Link rather than duplicate data sources
When we create a new minor version for a measure, or copy one to make a
new measure, we should create new links to the existing data sources
rather than copying the data sources verbatim and creating new records.
This is part of ensuring that each semantic data sources is represented
by only one record in our database.

 ## Ticket
https://trello.com/c/rXaKvG1N