Update source description for downloaded charts
On the main measure page, next to each chart's source, we show the
primary data source's title. However, in the downloaded version of the
chart, we currently just show the data source publisher's name. We
should be consistent about what we show. Let's show the data source
title in both cases.