Merge data source views
Combine the GET and POST for the data source views, reducing duplicated
code and potential for validation to differ. We also add the original
list of data sources to the POSTed route in the form, so that validation
can occur on submit.

Also updates the typing on the data source form, so that data source IDs
are always integers - we had a bug where some IDs were stored as strings
but compared against integers.