Update information on merge data source page
It's useful for analysts to know which data sources are associated to
published measures, as this implies they've already been "signed off".
This will give weight to preferring these as canonical data sources and
reduce the risk of accepting a data source from the testing space, or
that otherwise isn't finished yet.